backend/requirements.txt

Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
mysql-connector-python==8.1.0
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.6
